move icons too
