--- src/components/icons/arrowDownM.tsx
@@ -1,22 +1,0 @@
-import React from "react"
-
-const ArrowDownM = (props) => (
-  <svg
-    width="32"
-    height="32"
-    viewBox="0 0 32 32"
-    fill="none"
-    xmlns="http://www.w3.org/2000/svg"
-    {...props}
-  >
-    <path
-      d="M10.6667 13.3333L16 18.6666L21.3333 13.3333"
-      stroke="currentColor"
-      strokeWidth="2"
-      strokeLinecap="round"
-      strokeLinejoin="round"
-    />
-  </svg>
-)
-
-export default ArrowDownM
