--- src/components/icons/arrowDownMCrop.tsx
@@ -1,15 +1,0 @@
-import React from "react"
-
-const ArrowDownMCrop = (props) => (
-  <svg viewBox="0 0 32 32" {...props}>
-    <path
-      d="M22 13l-6 6-6-6"
-      fill="none"
-      stroke="currentColor"
-      strokeLinecap="round"
-      strokeWidth="1.8"
-    />
-  </svg>
-)
-
-export default ArrowDownMCrop
