--- src/components/icons/close.tsx
@@ -1,18 +1,0 @@
-import React, { FC } from "react"
-
-const Close: FC = () => {
-  return (
-    <svg xmlns="http://www.w3.org/2000/svg" width="32" height="32">
-      <path
-        d="M11 11l9.428 9.428m0-9.428L11 20.428"
-        fill="none"
-        stroke="#ffffff"
-        strokeLinecap="round"
-        strokeLinejoin="round"
-        strokeWidth="1.8"
-      />
-    </svg>
-  )
-}
-
-export default Close
