--- src/components/icons/closeS.tsx
@@ -1,23 +1,0 @@
-import React, { FC } from "react"
-
-const CloseS: FC = () => (
-  <svg
-    width="24"
-    height="24"
-    viewBox="0 0 24 24"
-    xmlns="http://www.w3.org/2000/svg"
-  >
-    <g
-      stroke="#232A36"
-      strokeWidth="1.2"
-      fill="none"
-      fillRule="evenodd"
-      strokeLinecap="round"
-      strokeLinejoin="round"
-    >
-      <path d="M7 7l10 10M17 7L7 17" />
-    </g>
-  </svg>
-)
-
-export default CloseS
