--- src/components/icons/info.tsx
@@ -1,18 +1,0 @@
-import React, { FC } from "react"
-
-const Info: FC = () => (
-  <svg
-    width="24"
-    height="24"
-    viewBox="4 4 28 28"
-    xmlns="http://www.w3.org/2000/svg"
-  >
-    <path
-      d="M16 7a9 9 0 018.996 8.735L25 16a9 9 0 01-8.735 8.996L16 25a9 9 0 01-9-9 9 9 0 018.735-8.996L16 7zm1 8h-2v6h2v-6zm0-4h-2v2h2v-2z"
-      fill="#3696B9"
-      fillRule="evenodd"
-    />
-  </svg>
-)
-
-export default Info
