--- src/components/icons/logoRedWhite.tsx
@@ -1,21 +1,0 @@
-import React, { FC } from "react"
-const LogoRedWhite: FC = () => {
-  return (
-    <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 243.04 37.13">
-      <path
-        fill="#F73B47"
-        d="M28.4 19.71H8.56a8.73 8.73 0 1 0 8.92 8.73v-1.73a1.46 1.46 0 0 0-2.92 0v1.67a5.83 5.83 0 1 1-5.83-5.83h10.93v5.83a8.74 8.74 0 1 0 8.74-8.67zm0 14.56a5.83 5.83 0 0 1-5.83-5.82v-5.88h5.83a5.83 5.83 0 0 1 0 11.65zM8.74 17.47H28.4a8.74 8.74 0 1 0-8.74-8.76v1.7a1.46 1.46 0 0 0 2.91 0v-1.7a5.83 5.83 0 1 1 5.83 5.82H17.48V8.71a8.74 8.74 0 1 0-8.74 8.73zm0-14.56a5.83 5.83 0 0 1 5.82 5.8v5.82H8.74a5.83 5.83 0 1 1 0-11.65z"
-      />
-      <path
-        fill="#FFF"
-        d="M57.03 7.97a9.74 9.74 0 0 1 6.8 2.61.52.52 0 0 1 0 .79l-1.23 1.26c-.2.25-.45.25-.71 0a7.62 7.62 0 0 0-4.8-1.92 7.29 7.29 0 0 0 0 14.58 7.52 7.52 0 0 0 4.83-1.85.52.52 0 0 1 .68 0l1.32 1.27a.53.53 0 0 1 0 .76 9.61 9.61 0 0 1-6.89 2.73 10.12 10.12 0 1 1 0-20.23zM68.03 27.19l8.62-18.91a.67.67 0 0 1 .48-.31h.28a.65.65 0 0 1 .48.31l8.57 18.91a.49.49 0 0 1-.48.73h-1.75a.65.65 0 0 1-.64-.42l-1.75-3.85h-9.18c-.57 1.29-1.16 2.55-1.72 3.85a.71.71 0 0 1-.64.42h-1.77a.49.49 0 0 1-.5-.73zm12.73-6l-3.46-7.7h-.14l-3.43 7.7zM92.03 8.78a.53.53 0 0 1 .53-.53h7.64a6.07 6.07 0 0 1 6.13 6 6.25 6.25 0 0 1-4.16 5.71l3.86 7.14a.53.53 0 0 1-.48.82h-2.19a.53.53 0 0 1-.45-.25l-3.74-7.45h-4.24v7.16a.55.55 0 0 1-.53.54h-1.83a.53.53 0 0 1-.53-.54zm8 9a3.49 3.49 0 0 0 3.43-3.48 3.45 3.45 0 0 0-3.43-3.35h-5v6.83zM116.46 8.78a.54.54 0 0 1 .54-.53h11.24a.53.53 0 0 1 .53.53v1.58a.53.53 0 0 1-.53.53h-8.88v6.49h7.5a.55.55 0 0 1 .53.53v1.55a.55.55 0 0 1-.53.53h-7.5v7.39a.56.56 0 0 1-.54.54h-1.79a.54.54 0 0 1-.54-.54zM143.03 7.97a10.12 10.12 0 1 1-10.1 10.14 10.1 10.1 0 0 1 10.1-10.14zm0 17.42a7.31 7.31 0 1 0-7.27-7.28 7.31 7.31 0 0 0 7.27 7.28zM159.03 8.78a.53.53 0 0 1 .53-.53h7.65a6.06 6.06 0 0 1 6.12 6 6.24 6.24 0 0 1-4.16 5.71l3.86 7.14a.53.53 0 0 1-.47.82h-2.2a.53.53 0 0 1-.45-.25l-3.73-7.45h-4.25v7.16a.55.55 0 0 1-.53.54h-1.87a.53.53 0 0 1-.53-.54zm7.95 9a3.49 3.49 0 0 0 3.43-3.48 3.45 3.45 0 0 0-3.43-3.35h-5v6.83z"
-      />
-      <path
-        fill="#F73B47"
-        d="M189.52 18.45l-6.57-9.39a.52.52 0 0 1 .45-.81h2.08a.58.58 0 0 1 .45.25l5.1 7.21 5.09-7.17a.6.6 0 0 1 .45-.25h2.11a.51.51 0 0 1 .44.81l-6.65 9.36v9a.56.56 0 0 1-.54.54h-1.9a.54.54 0 0 1-.54-.54zM212.03 7.97a10.12 10.12 0 1 1-10.12 10.14 10.1 10.1 0 0 1 10.12-10.14zm0 17.42a7.31 7.31 0 1 0-7.28-7.28 7.31 7.31 0 0 0 7.28 7.28zM227.76 8.78a.55.55 0 0 1 .53-.53h1.89a.53.53 0 0 1 .53.53v11.61c0 2.81 1.8 5 4.67 5a4.64 4.64 0 0 0 4.72-5V8.78a.53.53 0 0 1 .53-.53h1.88a.56.56 0 0 1 .54.53v11.8a7.65 7.65 0 0 1-15.29 0z"
-      />
-    </svg>
-  )
-}
-
-export default LogoRedWhite
