--- src/components/icons/notification.tsx
@@ -1,16 +1,0 @@
-import React from "react"
-
-const Notification = (props) => (
-  <svg
-    width="14"
-    height="14"
-    viewBox="0 0 14 14"
-    fill="none"
-    xmlns="http://www.w3.org/2000/svg"
-    {...props}
-  >
-    <circle cx="7" cy="7" r="7" fill="#F73B47" />
-  </svg>
-)
-
-export default Notification
