--- src/components/icons/ribbon.tsx
@@ -1,14 +1,0 @@
-import React from "react"
-
-const RibbonIcon = (props) => {
-  return (
-    <svg viewBox="0 0 130 47" fill="none" {...props}>
-      <path
-        d="M6.555 0S0 1.273 0 7.698V47s.728-7.917 6.555-7.917H130l-12.589-20.238L130 0H6.555z"
-        fill="currentColor"
-      />
-    </svg>
-  )
-}
-
-export default RibbonIcon
