--- src/components/icons/uploadImage.tsx
@@ -1,23 +1,0 @@
-import React from "react"
-
-const UploadImagesIcon = (props) => (
-  <svg viewBox="0 0 72 72" fill="none" {...props}>
-    <path
-      d="M44.638 30.503c4.67 4.672 4.67 12.245 0 16.916-4.671 4.671-12.245 4.671-16.916 0-4.67-4.671-4.67-12.244 0-16.915 4.671-4.672 12.245-4.672 16.916 0"
-      stroke="currentColor"
-      strokeWidth="3"
-      strokeLinecap="round"
-      strokeLinejoin="round"
-    />
-    <path
-      clipRule="evenodd"
-      d="M63 27v27c0 3.315-2.685 6-6 6H15c-3.315 0-6-2.685-6-6V27c0-3.315 2.685-6 6-6h6l4.386-7.512A3 3 0 0127.978 12H43.89c1.053 0 2.028.552 2.571 1.452L51 21h6c3.315 0 6 2.685 6 6z"
-      stroke="currentColor"
-      strokeWidth="3"
-      strokeLinecap="round"
-      strokeLinejoin="round"
-    />
-  </svg>
-)
-
-export default UploadImagesIcon
