+++ src/index.ts
@@ -21,10 +21,10 @@ import MenuItem from "./components/menu/item"
 import MenuElement, { MenuElementProps } from "./components/menu/element"
 import Intercom from "./components/intercom/index"
 import Input from "./components/input/index"
-import Profile from "./components/icons/profile"
-import LogoWhite from "./components/icons/logoWhite"
-import LogoRedWhite from "./components/icons/logoRedWhite"
-import LogoRedBlack from "./components/icons/logoRedBlack"
+import Profile from "./assets/dist/icons/profile"
+import LogoWhite from "./assets/dist/icons/logoWhite"
+import LogoRedWhite from "./assets/dist/icons/logoRedWhite"
+import LogoRedBlack from "./assets/dist/icons/logoRedBlack"
 import InputFilter from "./components/filters/input"
 import CheckboxFilter from "./components/filters/checkbox"
 import WithValidationError from "./components/fieldHelpers/withValidationError"
