+++ src/stories/filters/checkbox.stories.tsx
@@ -3,7 +3,7 @@ import React, { useState } from "react"
 import { action } from "@storybook/addon-actions"
 
 import StoryContainer from "../storyContainer"
-import ClockOutlinedTealIcon from "../../components/icons/clockOutlinedTeal"
+import ClockOutlinedTealIcon from "../../assets/src/icons/clockOutlinedTeal"
 import CheckboxFilter from "../../components/filters/checkbox"
 
 const options = [
